Log time to generate build script
- Move default log listener into its own library and make it public
- Set up logging in the entrypoint
- Surround generating the build script with `logTimedAsync`
