+++ build_runner/lib/src/build_script_generate/build_script_generate.dart
@@ -9,17 +9,21 @@ import 'package:build_config/build_config.dart';
 import 'package:build_runner/build_runner.dart';
 import 'package:code_builder/code_builder.dart';
 import 'package:dart_style/dart_style.dart';
+import 'package:logging/logging.dart';
 
+import '../logging/logging.dart';
 import 'types.dart' as types;
 
 const scriptLocation = '.dart_tool/build/build.dart';
 
 Future<Null> ensureBuildScript() async {
+  var log = new Logger('ensureBuildScript');
   var scriptFile = new io.File(scriptLocation);
   // TODO - how can we invalidate this?
   //if (scriptFile.existsSync()) return;
   scriptFile.createSync(recursive: true);
-  await scriptFile.writeAsString(await _generateBuildScript());
+  await logTimedAsync(log, 'Generating build script',
+      () async => scriptFile.writeAsString(await _generateBuildScript()));
 }
 
 Future<String> _generateBuildScript() async {
