+++ build_runner/lib/src/generate/options.dart
@@ -2,15 +2,13 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 import 'dart:async';
-import 'dart:convert';
-import 'dart:io';
 
 import 'package:logging/logging.dart';
-import 'package:stack_trace/stack_trace.dart';
 
 import '../asset/file_based.dart';
 import '../asset/reader.dart';
 import '../asset/writer.dart';
+import '../logging/std_io_logging.dart';
 import '../package_graph/package_graph.dart';
 import 'directory_watcher_factory.dart';
 
@@ -51,7 +49,7 @@ class BuildOptions {
     /// Set up logging
     logLevel ??= Level.INFO;
     Logger.root.level = logLevel;
-    logListener = Logger.root.onRecord.listen(onLog ?? _defaultLogListener);
+    logListener = Logger.root.onRecord.listen(onLog ?? stdIOLogListener);
 
     /// Set up other defaults.
     debounceDelay ??= const Duration(milliseconds: 250);
@@ -62,56 +60,5 @@ class BuildOptions {
     deleteFilesByDefault ??= writeToCache ?? false;
     writeToCache ??= false;
     skipBuildScriptCheck ??= false;
-  }
-}
-
-final _cyan = _isPosixTerminal ? '\u001b[36m' : '';
-final _yellow = _isPosixTerminal ? '\u001b[33m' : '';
-final _red = _isPosixTerminal ? '\u001b[31m' : '';
-final _endColor = _isPosixTerminal ? '\u001b[0m' : '';
-final _isPosixTerminal =
-    !Platform.isWindows && stdioType(stdout) == StdioType.TERMINAL;
-
-void _defaultLogListener(LogRecord record) {
-  var color;
-  if (record.level < Level.WARNING) {
-    color = _cyan;
-  } else if (record.level < Level.SEVERE) {
-    color = _yellow;
-  } else {
-    color = _red;
-  }
-  var header = '${_isPosixTerminal ? '\x1b[2K\r' : ''}'
-      '$color[${record.level}]$_endColor ${record.loggerName}: '
-      '${record.message}';
-  var lines = <Object>[header];
-
-  if (record.error != null) {
-    lines.add(record.error);
-  }
-
-  if (record.stackTrace != null) {
-    if (record.stackTrace is Trace) {
-      lines.add((record.stackTrace as Trace).terse);
-    } else {
-      lines.add(record.stackTrace);
-    }
-  }
-
-  var message = new StringBuffer(lines.join('\n'));
-
-  // We always add an extra newline at the end of each message, so it
-  // isn't multiline unless we see > 2 lines.
-  var multiLine = LineSplitter.split(message.toString()).length > 2;
-
-  if (record.level > Level.INFO || !_isPosixTerminal || multiLine) {
-    // Add an extra line to the output so the last line isn't written over.
-    message.writeln('');
-  }
-
-  if (record.level >= Level.SEVERE) {
-    stderr.write(message);
-  } else {
-    stdout.write(message);
   }
 }
